PCI: endpoint: nvme: Assign ofst variable before use
The variable ofst is used to compute the number of segments
required for the PRP transfers but this variable was only set
after the computation of the number of segments. This lead to
incorrect computation of number of segments in some cases.
Assign ofst before it is referenced in the computation.

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>