PCI: endpoint: nvme: Check flag bit correctly
The variable sq_flags (u16) was checked directly against
a flag for equality but only one bit should be checked
for the flag. This was implemented correctly for the
cq_flags but not for the sq_flags. Check only the flag
bit and not the whole variable against the flag.

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>