PCI: endpoint: nvme: Only enable BAR_0
NVMe PCIe transport specifications Fig. 10 p. 16 specifies
that NVMe BAR0 is mandatory and BAR1 should not be used
unless is contains the upper 32-bits for BAR0 configured
as a 64-bit BAR.

BAR2,3,4, and 5 can be used for vendor specific functions.
BAR2 can be used for Index/Data pair register base address.

But they are not necessary in this NVMe controller so disable
them.

Signed-off-by: Rick Wertenbroek <rick.wertenbroek@gmail.com>